refactor(filters): extract TimeRangeSection from FilterBarClient
